Add Sonar analyzer

+++ src/Duracellko.PlanningPoker/Controllers/PlanningPokerController.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
@@ -70,6 +71,7 @@ namespace Duracellko.PlanningPoker.Controllers
         /// <summary>
         /// Gets a collection of Scrum team names.
         /// </summary>
+        [SuppressMessage("Critical Code Smell", "S2365:Properties should not make collection or array copies", Justification = "Creates copy to be thread-safe.")]
         public IEnumerable<string> ScrumTeamNames
         {
             get
@@ -308,17 +310,14 @@ namespace Duracellko.PlanningPoker.Controllers
 
             LogScrumTeamMessage(team, e.Message);
 
-            if (e.Message.MessageType == MessageType.MemberDisconnected)
+            if (e.Message.MessageType == MessageType.MemberDisconnected && !IsTeamActive(team))
             {
-                if (!IsTeamActive(team))
-                {
-                    saveTeam = false;
-                    OnTeamRemoved(team);
+                saveTeam = false;
+                OnTeamRemoved(team);
 
-                    _scrumTeams.TryRemove(team.Name, out _);
-                    Repository.DeleteScrumTeam(team.Name);
-                    _logger.ScrumTeamRemoved(team.Name);
-                }
+                _scrumTeams.TryRemove(team.Name, out _);
+                Repository.DeleteScrumTeam(team.Name);
+                _logger.ScrumTeamRemoved(team.Name);
             }
 
             if (saveTeam)
@@ -345,17 +344,8 @@ namespace Duracellko.PlanningPoker.Controllers
                     {
                         if (VerifyTeamActive(team))
                         {
-                            var teamLock = new object();
-                            result = new Tuple<ScrumTeam, object>(team, teamLock);
-                            if (_scrumTeams.TryAdd(team.Name, result))
-                            {
-                                OnTeamAdded(team);
-                            }
-                            else
-                            {
-                                result = null;
-                                retry = true;
-                            }
+                            result = AttachLoadedScrumTeam(team);
+                            retry = result == null;
                         }
                         else
                         {
@@ -371,6 +361,19 @@ namespace Duracellko.PlanningPoker.Controllers
         private void SaveScrumTeam(ScrumTeam team)
         {
             Repository.SaveScrumTeam(team);
+        }
+
+        private Tuple<ScrumTeam, object>? AttachLoadedScrumTeam(ScrumTeam team)
+        {
+            var teamLock = new object();
+            var result = new Tuple<ScrumTeam, object>(team, teamLock);
+            if (_scrumTeams.TryAdd(team.Name, result))
+            {
+                OnTeamAdded(team);
+                return result;
+            }
+
+            return null;
         }
 
         private bool VerifyTeamActive(ScrumTeam team)
